Resolve compilation failure on Ubuntu 18.04, 20.04
Ubuntu 17.04 made a behavior change in GCC, having GCC compile programs with PIE support by default. Since the libasncodec.a file that currently ships in the repository was compiled on a system without PIE enabled in GCC, builds fail on systems with the new GCC behavior. This commit explicitly disables PIE compilation in the S1AP component of nucleus due to this so that Nucleus can compile on systems with GCC configured to compile with PIE. Eventually the libasncodec.a file and other ASN1 files for S1AP need to be rebuilt with PIE support.